Get some initial infrastructure flushed out
No doubt all of this is going to change,
but I needed to get far enough to be able to make progress
and feel what is good and bad about this design.

iow: I don't know what "right" looks like, so I'll go
with "works" and then refactor to "right" when I can feel it.